First attempt at testing depth > 1 with repeat2/index2 args
This fails on the last assertion: the value is apparently `null`. I do not understand how that is possible!

The full test passes in Web Forms with `indexed-repeat` implemented as I thought I understood from the last round of clarification. Is there something I’m still missing? Is this a JR bug?

I did consider the possibility that there’s an issue with the order of “act” state changes and corresponding calculation updates, but I reordered the “act” steps in a way I would hope/expect to minimize that possibilty.